+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 - [#674](https://github.com/convertigo/convertigo/issues/674) [Studio] Fixed, hightlight elements supported for recursive shared components
 - [#675](https://github.com/convertigo/convertigo/issues/675) [Studio] Fixed, project creation invite to perform the registration wizard
 - [#676](https://github.com/convertigo/convertigo/issues/676) [Step] Fixed, JsonSource handles a single string element
+- [#678](https://github.com/convertigo/convertigo/issues/678) [Mac] Fixed, drag-n-drop from the NGX Palette to the Builder
 
 ---